requirements.txt
asgiref==3.2.10
Django==3.0.7
django-filter==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
importlib-metadata==1.6.1
Markdown==3.2.2
pkg-resources==0.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
zipp==3.1.0
